website_crm_partner: updated form

+++ website_crm_partner/res_partner.py
@@ -88,7 +88,7 @@ class website_crm_partner(http.Controller):
     MODULE_BASE_PATH + '<model("res.partner"):partner>',
     MODULE_BASE_PATH,
     MODULE_BASE_PATH + 'add',
-    MODULE_BASE_PATH + 'delete',
+    MODULE_BASE_PATH + '<model("res.partner"):partner>/delete',
     MODULE_BASE_PATH + '<model("res.partner"):partner>/edit',
     MODULE_BASE_PATH + '<string:search>/search',
     MODULE_BASE_PATH + 'set_login',
@@ -97,38 +97,40 @@ class website_crm_partner(http.Controller):
 
         search_domain = [('type','=','contact')]
         model = 'res.partner'
-        fields =  ['name','phone','email']
-        template = {'list': 'website_crm_partner.partner_list', 'detail': 'website_crm_partner.partner_detail'}
+        fields =  ['name','phone','email','comments']
+        template = {'list': 'website_crm_partner.object_list', 'detail': 'website_crm_partner.object_detail'}
 
 
         if request.httprequest.url[-4:] == 'edit': #Edit
             if request.httprequest.method == 'GET':
-                return request.render(template['detail'], {'model': model, 'partner': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'db': request.db, 'mode': 'edit'})
+                return request.render(template['detail'], {'model': model, 'object': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db, 'mode': 'edit'})
             else:
                 partner.write({
                     f: post.get(f) for f in fields
                 })
-                return request.render(template['detail'], {'model': model, 'partner': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'db': request.db, 'mode': 'view'})
+                return request.render(template['detail'], {'model': model, 'object': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db, 'mode': 'view'})
 
         elif request.httprequest.url[-3:] == 'add': #Add
             if request.httprequest.method == 'GET':
-                return request.render(template['detail'], {'model': model, 'partner': None, 'fields': fields, 'root': MODULE_BASE_PATH, 'db': request.db,'mode': 'edit'})
+                return request.render(template['detail'], {'model': model, 'object': None, 'fields': fields, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db,'mode': 'edit'})
             else:
                 record = { f: post.get(f) for f in fields }
                 record['type'] = 'contact'
                 partner = request.env['res.partner'].create(record)
-                return request.render(template['detail'], {'model': model, 'partner': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'db': request.db, 'mode': 'view'})
+                return request.render(template['detail'], {'model': model, 'object': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db, 'mode': 'view'})
         elif request.httprequest.url[-6:] == 'delete': #Delete
-            partner.unlink()
+            if partner:
+                partner.unlink()
         elif request.httprequest.url[-6:] == 'search': #Search
             if request.httprequest.method == 'POST':
                 search = post.get('search')
             search_domain.append(('name','like',search))
         elif partner:  # Detail
-            return request.render(template['detail'], {'model': model, 'partner': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'db': request.db, 'mode': 'view'})
+            return request.render(template['detail'], {'model': model, 'object': partner, 'fields': fields, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db, 'mode': 'view'})
 
         return request.render(template['list'], {
-            'partners': request.env['res.partner'].search(search_domain, order='name',limit=25),
+            'objects': request.env['res.partner'].search(search_domain, order='name',limit=25),
+            'title': MODULE_TITLE,
             'root': MODULE_BASE_PATH,
             'db': request.db,
         })
@@ -138,14 +140,14 @@ class website_crm_partner(http.Controller):
         if request.httprequest.url[-9:] == 'set_login': #set login form
             if request.httprequest.method == 'POST':
                 return werkzeug.utils.redirect(MODULE_BASE_PATH, 302)
-            return request.render('website_crm_partner.set_login', {'partner': None, 'root': MODULE_BASE_PATH, 'db': request.db,})
+            return request.render('website_crm_partner.set_login', {'partner': None, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db,})
 
 
 
     @http.route(['/mobile/security/<model("res.partner"):partner>', '/mobile/security'], type='http', auth="user", website=True)
     def mobile_security(self, partner=False, **post):
         partners = request.env['res.partner'].search([('type', '=', 'contact')], order='name',limit=25)
-        return request.render('website_crm_partner.partner_list', {'partners': partners, 'root': MODULE_BASE_PATH, 'db': request.db,})
+        return request.render('website_crm_partner.partner_list', {'partners': partners, 'root': MODULE_BASE_PATH, 'title': MODULE_TITLE, 'db': request.db,})
 
     #~ @http.route(['/allcategory/<model("product.category"):category>', ], type='http', auth="public", website=True)
     #~ def get_category(self, parent_id=1, **post):
